Setting up live server
The live server is a VS Code plugin that allows you to run your website/page on your localhost. The advantage of that is that you don't have to click on the page in the file system of the computer to run your website, nor do you have to reload the page to see the changes you made on it, the page will change automatically.