meta tags and search engines
I learned what meta tags are, how to use them and their attributes and how search engines react to those attributes. For instance, the attribute "description" displays a description for the website when it's searched on a search engine. Another nice meta tag attribute is the "robots" one, which allows the website to be hidden from searches on the internet.