images
Local and remote images can be added to an HTML page with the same tag: <img>. It has two main attributes: src and alt. If the image is local, place its directory in the src, if it's remote, its url goes in the same place. For the alt attribute, write whatever text you'd like to be displayed if the image can't be loaded. If your page is pure HTML, make sure to give the image some constraint with the width and height attributes, so the image is loaded correctly. Although, using a styling language such as CSS for doing that is more recommended.