tables
Tables have head, body, rows and columns. Inside the tag table, we place the tags head and body. In order to add content to each portion, we insert tr (table rows) and insert the data inside the tag th for the head, and td for the body. Each th and td adds one column to the table.